pyproject.toml
[tool.poetry]
name = "Backend"
version = "0.1.0"
description = ""
authors = ["reach bheru <reachbheru@gmail.com>"]
packages = [{include = "Backend"}]
package-mode = true

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = "^2.11.7"
gunicorn = "^23.0.0"
rich = "^14.1.0"
jinja2 = "^3.1.6"
langgraph = "^0.6.6"
motor = "^3.7.1"
langchain = "^0.3.27"
langchain-google-genai = "^2.1.10"
cloudinary = "^1.44.1"
pymongo = "^4.14.1"
google-cloud-vision = "^3.10.2"
python-multipart = "^0.0.20"
pillow = "^11.3.0"
boto3 = "^1.40.23"

[tool.poetry.scripts]
dev = "Backend.main:dev"
start = "Backend.main:start"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
